requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
anthropic>=0.40.0
sqlalchemy>=2.0.0
plotly>=5.18.0
streamlit>=1.35.0
fastapi>=0.111.0
uvicorn>=0.29.0
pandas>=2.2.0
scipy>=1.13.0
python-dotenv>=1.0.0
requests>=2.31.0
